fix(tests): fix tracer name detection
When running tests, the test runner determines the name of the tracer ('pure'
or 'native', based on the name of the tracer). This name is used to create a
proper 'pure' or 'native' subdirectory for the trace output. The tracer name
detection used to check for the presence of the substring 'native' in the tracer
name. Unfortunately, this didn't work at all, because the 'native' tracer
doesn't have 'native' in its name. The two tracers are called:
  1) codetracer-pure-ruby-recorder
  2) codetracer-ruby-recorder
So, both tracers were detected as 'pure' (not 'native'), and wrote their output
in the 'pure' subdirectory. This commit inverts the check, so now we test for
the presence of the word 'pure', instead of 'native'.