refactor: more use of cstr_to_string() + other simplifications
Don't expect a &[&str] parameter in the struct_value() function. In many cases,
it's not practical to require from the caller to build a vector of references.
It only works in very limited cases, in other cases, the caller has to jump
through hoops. On the other hand, requiring a &[String] is also limiting in many
cases, and breaks other callers. And you can't really do both, because Rust
doesn't have function overloading. That's why we now take a generic &[T]
parameter, where <T: AsRef<str> + ToString>.